requirements.txt
## The following requirements were added by pip --freeze:
Django==1.7
coverage==3.7.1
factory-boy==2.4.1
mock==1.0.1
pep8==1.5.7
psycopg2==2.5.4
pyflakes==0.8.1
requests==2.4.3
selenium==2.43.0
wsgiref==0.1.2
